refactor: dedupe `CallRequest`/`TransactionRequest`
