+++ crates/rpc-types/src/eth/call.rs
@@ -1,5 +1,5 @@
-use crate::{AccessList, BlockId, BlockOverrides};
-use alloy_primitives::{Address, Bytes, B256, U256, U64, U8};
+use crate::{request::TransactionRequest, BlockId, BlockOverrides};
+use alloy_primitives::Bytes;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
 /// Bundle of transactions
@@ -7,7 +7,7 @@ use serde::{Deserialize, Deserializer, Serialize, Serializer};
 #[serde(default, rename_all = "camelCase")]
 pub struct Bundle {
     /// All transactions to execute
-    pub transactions: Vec<CallRequest>,
+    pub transactions: Vec<TransactionRequest>,
     /// Block overrides to apply
     pub block_override: Option<BlockOverrides>,
 }
@@ -97,201 +97,6 @@ impl<'de> Deserialize<'de> for TransactionIndex {
     }
 }
 
-/// Call request for `eth_call` and adjacent methods.
-#[derive(Debug, Clone, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
-#[serde(default, rename_all = "camelCase")]
-pub struct CallRequest {
-    /// From
-    pub from: Option<Address>,
-    /// To
-    pub to: Option<Address>,
-    /// Gas Price
-    pub gas_price: Option<U256>,
-    /// EIP-1559 Max base fee the caller is willing to pay
-    pub max_fee_per_gas: Option<U256>,
-    /// EIP-1559 Priority fee the caller is paying to the block author
-    pub max_priority_fee_per_gas: Option<U256>,
-    /// Gas
-    pub gas: Option<U256>,
-    /// Value
-    pub value: Option<U256>,
-    /// Transaction input data
-    #[serde(default, flatten)]
-    pub input: CallInput,
-    /// Nonce
-    pub nonce: Option<U64>,
-    /// chain id
-    pub chain_id: Option<U64>,
-    /// AccessList
-    pub access_list: Option<AccessList>,
-    /// Max Fee per Blob gas for EIP-4844 transactions
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_fee_per_blob_gas: Option<U256>,
-    /// Blob Versioned Hashes for EIP-4844 transactions
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blob_versioned_hashes: Option<Vec<B256>>,
-    /// EIP-2718 type
-    #[serde(rename = "type", skip_serializing_if = "Option::is_none")]
-    pub transaction_type: Option<U8>,
-}
-
-impl CallRequest {
-    /// Returns the configured fee cap, if any.
-    ///
-    /// The returns `gas_price` (legacy) if set or `max_fee_per_gas` (EIP1559)
-    #[inline]
-    pub fn fee_cap(&self) -> Option<U256> {
-        self.gas_price.or(self.max_fee_per_gas)
-    }
-
-    /// Returns true if the request has a `blobVersionedHashes` field but it is empty.
-    #[inline]
-    pub fn has_empty_blob_hashes(&self) -> bool {
-        self.blob_versioned_hashes.as_ref().map(|blobs| blobs.is_empty()).unwrap_or(false)
-    }
-
-    /// Sets the `from` field in the call to the provided address
-    #[inline]
-    pub const fn from(mut self, from: Address) -> Self {
-        self.from = Some(from);
-        self
-    }
-
-    /// Sets the `to` field in the call to the provided address
-    #[inline]
-    pub const fn to(mut self, to: Option<Address>) -> Self {
-        self.to = to;
-        self
-    }
-
-    /// Sets the `gas` field in the transaction to the provided value
-    pub const fn gas(mut self, gas: U256) -> Self {
-        self.gas = Some(gas);
-        self
-    }
-
-    /// Sets the `gas_price` field in the transaction to the provided value
-    /// If the internal transaction is an EIP-1559 one, then it sets both
-    /// `max_fee_per_gas` and `max_priority_fee_per_gas` to the same value
-    pub const fn gas_price(mut self, gas_price: U256) -> Self {
-        // todo: Add legacy support
-        self.max_fee_per_gas = Some(gas_price);
-        self.max_fee_per_gas = Some(gas_price);
-        self
-    }
-
-    /// Sets the `value` field in the transaction to the provided value
-    pub const fn value(mut self, value: U256) -> Self {
-        self.value = Some(value);
-        self
-    }
-
-    /// Sets the `nonce` field in the transaction to the provided value
-    pub const fn nonce(mut self, nonce: U64) -> Self {
-        self.nonce = Some(nonce);
-        self
-    }
-
-    /// Calculates the address that will be created by the transaction, if any.
-    ///
-    /// Returns `None` if the transaction is not a contract creation (the `to` field is set), or if
-    /// the `from` or `nonce` fields are not set.
-    pub fn calculate_create_address(&self) -> Option<Address> {
-        if self.to.is_some() {
-            return None;
-        }
-        let from = self.from.as_ref()?;
-        let nonce = self.nonce?;
-        Some(from.create(nonce.to()))
-    }
-}
-
-/// Helper type that supports both `data` and `input` fields that map to transaction input data.
-///
-/// This is done for compatibility reasons where older implementations used `data` instead of the
-/// newer, recommended `input` field.
-///
-/// If both fields are set, it is expected that they contain the same value, otherwise an error is
-/// returned.
-#[derive(Debug, Clone, Default, Eq, PartialEq, Serialize, Deserialize, Hash)]
-pub struct CallInput {
-    /// Transaction data
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub input: Option<Bytes>,
-    /// Transaction data
-    ///
-    /// This is the same as `input` but is used for backwards compatibility: <https://github.com/ethereum/go-ethereum/issues/15628>
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<Bytes>,
-}
-
-impl CallInput {
-    /// Creates a new instance with the given input data.
-    pub const fn new(data: Bytes) -> Self {
-        Self::maybe_input(Some(data))
-    }
-
-    /// Creates a new instance with the given input data.
-    pub const fn maybe_input(input: Option<Bytes>) -> Self {
-        Self { input, data: None }
-    }
-
-    /// Consumes the type and returns the optional input data.
-    #[inline]
-    pub fn into_input(self) -> Option<Bytes> {
-        self.input.or(self.data)
-    }
-
-    /// Consumes the type and returns the optional input data.
-    ///
-    /// Returns an error if both `data` and `input` fields are set and not equal.
-    #[inline]
-    pub fn try_into_unique_input(self) -> Result<Option<Bytes>, CallInputError> {
-        self.check_unique_input().map(|()| self.into_input())
-    }
-
-    /// Returns the optional input data.
-    #[inline]
-    pub fn input(&self) -> Option<&Bytes> {
-        self.input.as_ref().or(self.data.as_ref())
-    }
-
-    /// Returns the optional input data.
-    ///
-    /// Returns an error if both `data` and `input` fields are set and not equal.
-    #[inline]
-    pub fn unique_input(&self) -> Result<Option<&Bytes>, CallInputError> {
-        self.check_unique_input().map(|()| self.input())
-    }
-
-    fn check_unique_input(&self) -> Result<(), CallInputError> {
-        if let (Some(input), Some(data)) = (&self.input, &self.data) {
-            if input != data {
-                return Err(CallInputError::default());
-            }
-        }
-        Ok(())
-    }
-}
-
-impl From<Bytes> for CallInput {
-    fn from(input: Bytes) -> Self {
-        Self { input: Some(input), data: None }
-    }
-}
-
-impl From<Option<Bytes>> for CallInput {
-    fn from(input: Option<Bytes>) -> Self {
-        Self { input, data: None }
-    }
-}
-
-/// Error thrown when both `data` and `input` fields are set and not equal.
-#[derive(Debug, Default, thiserror::Error)]
-#[error("both \"data\" and \"input\" are set and not equal. Please use \"input\" to pass transaction call data")]
-#[non_exhaustive]
-pub struct CallInputError;
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -309,30 +114,5 @@ mod tests {
         let s = "-2";
         let res = serde_json::from_str::<TransactionIndex>(s);
         assert!(res.is_err());
-    }
-
-    #[test]
-    fn serde_call_request() {
-        let s = r#"{"accessList":[],"data":"0x0902f1ac","to":"0xa478c2975ab1ea89e8196811f51a7b7ade33eb11","type":"0x02"}"#;
-        let _req = serde_json::from_str::<CallRequest>(s).unwrap();
-    }
-
-    #[test]
-    fn serde_unique_call_input() {
-        let s = r#"{"accessList":[],"data":"0x0902f1ac", "input":"0x0902f1ac","to":"0xa478c2975ab1ea89e8196811f51a7b7ade33eb11","type":"0x02"}"#;
-        let req = serde_json::from_str::<CallRequest>(s).unwrap();
-        assert!(req.input.try_into_unique_input().unwrap().is_some());
-
-        let s = r#"{"accessList":[],"data":"0x0902f1ac","to":"0xa478c2975ab1ea89e8196811f51a7b7ade33eb11","type":"0x02"}"#;
-        let req = serde_json::from_str::<CallRequest>(s).unwrap();
-        assert!(req.input.try_into_unique_input().unwrap().is_some());
-
-        let s = r#"{"accessList":[],"input":"0x0902f1ac","to":"0xa478c2975ab1ea89e8196811f51a7b7ade33eb11","type":"0x02"}"#;
-        let req = serde_json::from_str::<CallRequest>(s).unwrap();
-        assert!(req.input.try_into_unique_input().unwrap().is_some());
-
-        let s = r#"{"accessList":[],"data":"0x0902f1ac", "input":"0x0902f1","to":"0xa478c2975ab1ea89e8196811f51a7b7ade33eb11","type":"0x02"}"#;
-        let req = serde_json::from_str::<CallRequest>(s).unwrap();
-        assert!(req.input.try_into_unique_input().is_err());
     }
 }
